[Feat] extract non-custodial process to a global hook.
